requirements.txt
aiofiles==22.1.0
asttokens==2.1.0
backcall==0.2.0
decorator==5.1.1
executing==1.2.0
flake8==5.0.4
ipdb==0.13.9
ipython==8.6.0
jedi==0.18.1
matplotlib-inline==0.1.6
mccabe==0.7.0
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.32
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pytesseract==0.3.10
six==1.16.0
stack-data==0.6.0
toml==0.10.2
traitlets==5.5.0
wcwidth==0.2.5
